Address Sonar findings on the lifecycle interception change
LifecycleCallbackMethod overrides equals and hashCode alongside its
superclass, since it adds the callback it stands for. The chain no longer
guards parameters that are never null and takes the bean context from
the resolution context, which keeps doIntercept under the parameter
limit and drops the parameter from the new overloads. Constant indexes
are boxed without an explicit unboxing.

+++ core-processor/src/main/java/io/micronaut/inject/writer/ExecutableMethodsDefinitionWriter.java
@@ -498,7 +498,7 @@ public class ExecutableMethodsDefinitionWriter implements Buildable<OutputObject
         if (indexes.isEmpty()) {
             return null;
         }
-        List<ExpressionDef> values = indexes.stream().map(index -> (ExpressionDef) TypeDef.Primitive.INT.constant(index.intValue())).toList();
+        List<ExpressionDef> values = indexes.stream().map(index -> (ExpressionDef) TypeDef.Primitive.INT.constant(index)).toList();
         return MethodDef.override(method)
             .build((aThis, methodParameters) -> TypeDef.Primitive.INT.array().instantiate(values).returning());
     }
